Get binary parsing working.

+++ examples/log_threads/log_threads.cpp
@@ -1,20 +1,18 @@
-#include <string>
-#include <thread>
-
+#include <min_logger/min_logger.h>
 #include <pthread.h>
 
-#include <min_logger/min_logger.h>
+#include <string>
+#include <thread>
 
 using namespace std;
 
 // The function we want to execute on the new thread.
-void task(string msg)
-{
+void task(string msg) {
     auto handle = pthread_self();
     msg = string("task") + msg;
     pthread_setname_np(handle, (msg).c_str());
 
-    for (int i=0; i < 5; i++) {
+    for (uint64_t i = 0; i < 5; i++) {
         MIN_LOGGER_ENTER(MIN_LOGGER_DEBUG, "TASK_LOOP");
         MIN_LOGGER_RECORD_STRING(MIN_LOGGER_INFO, "T_NAME", msg.c_str());
         MIN_LOGGER_RECORD_U64(MIN_LOGGER_INFO, "LOOP_COUNT", i);
@@ -24,15 +22,15 @@ void task(string msg)
     }
 }
 
-int main()
-{
+int main() {
     min_logger_write_thread_names();
 
     // Constructs the new thread and runs it. Does not block execution.
     thread t1(task, "1");
     thread t2(task, "2");
 
-    // Makes the main thread wait for the new thread to finish execution, therefore blocks its own execution.
+    // Makes the main thread wait for the new thread to finish execution, therefore blocks its own
+    // execution.
     t1.join();
     t2.join();
 
@@ -42,7 +40,8 @@ int main()
     thread t3(task, "3");
     thread t4(task, "4");
 
-    // Makes the main thread wait for the new thread to finish execution, therefore blocks its own execution.
+    // Makes the main thread wait for the new thread to finish execution, therefore blocks its own
+    // execution.
     t3.join();
     t4.join();
 }
